[Services] Allow for docker paths and system packages
Summary:
This allows for later diffs to not fail CI

This handles the existing and later workflows.

This work will be reverted later

Depends on D4685

Test Plan: Blob, backup, and tunnelbroker CI gates pass

Reviewers: atul, abosh, max, karol

Subscribers: ashoat, tomek, adrian

Differential Revision: https://phab.comm.dev/D4734

+++ services/lib/docker/build_service.sh
@@ -2,24 +2,33 @@
 
 set -e
 
+# Allow scripts to be called from anywhere
+SCRIPT_DIR=$(cd "$(dirname "$0")"; pwd -P)
+
 # folly hack - https://github.com/facebook/folly/pull/1231
-sed -i 's/#if __has_include(<demangle.h>)/#if __has_include(<Demangle.h>)/g' /usr/lib/folly/folly/detail/Demangle.h
+if [[ -f /usr/lib/folly/folly/detail/Demangle.h ]]; then
+  # Legacy path-centric workflow
+  sed -i 's/#if __has_include(<demangle.h>)/#if __has_include(<Demangle.h>)/g' \
+    /usr/lib/folly/folly/detail/Demangle.h
+  rm -rf lib
+  mkdir lib
+  pushd lib
+  ln -s /usr/lib/folly .
+  ln -s /usr/lib/glog .
+  ln -s /usr/lib/double-conversion .
+  popd # lib
 
-rm -rf lib
-mkdir lib
-pushd lib
-ln -s /usr/lib/folly .
-ln -s /usr/lib/glog .
-ln -s /usr/lib/double-conversion .
-popd # lib
+  rm -rf _generated
+  mkdir _generated
 
-rm -rf _generated
-mkdir _generated
+  "${SCRIPT_DIR}"/proto_codegen.sh
+else
+  # New installation centric workflow
+  sed -i 's/#if __has_include(<demangle.h>)/#if __has_include(<Demangle.h>)/g' \
+    /usr/local/include/folly/detail/Demangle.h
+fi
 
 rm -rf cmake/build
 mkdir -p cmake/build
 
-# Allow scripts to be called from anywhere
-SCRIPT_DIR=$(cd "$(dirname "$0")"; pwd -P)
-"${SCRIPT_DIR}"/proto_codegen.sh
 "${SCRIPT_DIR}"/build_sources.sh
